% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_folder_id}
\alias{derive_folder_id}
\title{Derive Google Drive Folder ID}
\usage{
derive_folder_id(link)
}
\arguments{
\item{link}{URL of Google Drive Folder}
}
\value{
A character vector
}
\description{
Derive Google Drive Folder ID
}
\examples{
x = "https://drive.google.com/drive/folders/1pXBQQdd1peI56GtQT-jEZ59xSmhqQlFC?usp=sharing"
derive_folder_id(x)
x = "1pXBQQdd1peI56GtQT-jEZ59xSmhqQlFC"
derive_folder_id(x)
}
