% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_presentation_id}
\alias{derive_presentation_id}
\title{Derive the ID of a presentation from the URL}
\usage{
derive_presentation_id(link)
}
\arguments{
\item{link}{URL of Google Slides presentation}
}
\value{
A character vector
}
\description{
Derive the ID of a presentation from the URL
}
\examples{
link = paste0("https://docs.google.com/presentation/d/",
"1Tg-GTGnUPduOtZKYuMoelqUNZnUp3vvg_7TtpUPL7e8",
"/edit#slide=id.g154aa4fae2_0_58")
derive_presentation_id(link)
}
