% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download Google Slides}
\usage{
download(link, type = "pptx")
}
\arguments{
\item{link}{Link to Google Slides presentation}

\item{type}{Type of file to download as output. Usually
\code{pdf} or \code{pptx}}
}
\value{
Downloaded file (in temporary directory)
}
\description{
Download Google Slides
}
\note{
This downloads presentations if they are public and also try to make
sure it does not fail on large files
}
