% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsplyr-package.R
\docType{package}
\name{gsplyr-package}
\alias{gsplyr}
\alias{gsplyr-package}
\title{gsplyr: Google Slides Manipulation}
\description{
The goal of gsplyr is to provide a set of functions designed for the manipulation of Google Slides.
}
\author{
\strong{Maintainer}: Howard Baek \email{howardbaek.fh@gmail.com} (\href{https://orcid.org/009-0000-8942-1618}{ORCID})

}
\keyword{internal}
